--- surf-tab-api/src/main/kotlin/dev/slne/surf/tab/api/model/TabGroup.kt
@@ -1,10 +1,0 @@
-package dev.slne.surf.tab.api.model
-
-import it.unimi.dsi.fastutil.objects.ObjectSet
-
-interface TabGroup {
-  val name: String
-  val display: String
-
-  val servers: ObjectSet<TabServer>
-}
