+++ surf-tab-api/src/main/kotlin/dev/slne/surf/tab/api/server/TabGroup.kt
@@ -1,0 +1,10 @@
+package dev.slne.surf.tab.api.server
+
+import it.unimi.dsi.fastutil.objects.ObjectSet
+
+interface TabGroup {
+    val name: String
+    val display: String
+
+    val servers: ObjectSet<TabServer>
+}
