+++ surf-tab-core/src/main/kotlin/dev/slne/surf/tab/core/model/TabEntryImpl.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.tab.core.model
 
-import dev.slne.surf.tab.api.model.TabEntry
-import dev.slne.surf.tab.api.model.TabGameMode
+import dev.slne.surf.tab.api.TabEntry
+import dev.slne.surf.tab.api.player.TabGameMode
 import net.kyori.adventure.text.Component
 import java.util.*
