+++ surf-tab-velocity/src/main/kotlin/dev/slne/surf/tab/velocity/util/player-util.kt
@@ -2,7 +2,7 @@ package dev.slne.surf.tab.velocity.util
 
 import com.github.retrooper.packetevents.protocol.player.GameMode
 import com.velocitypowered.api.proxy.Player
-import dev.slne.surf.tab.api.model.TabGameMode
+import dev.slne.surf.tab.api.player.TabGameMode
 import dev.slne.surf.tab.api.player.TabPlayer
 import dev.slne.surf.tab.core.factory.tabPlayerFactory
 import dev.slne.surf.tab.velocity.plugin
